+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+# Unreleased
+
+- Support absolute and relative toolchain paths in `lean-toolchain` etc.
+
 # 4.1.2 - 2025-05-26
 
 - Ignore errors when writing to `.elan/known-projects`, e.g. when it is read-only.